+++ HISTORY.md
@@ -2,6 +2,7 @@ Work in progress
 ----------------
 ### Fixes
 - Fix position mapping for a match at the first character in a file.
+- Make `yalafi.shell` subprocess cancellable.
 
 
 Version 0.0.1 (2023/03/08)